+++ cores/STC/stc_usb_service.c
@@ -1,0 +1,3 @@
+/* SPDX-License-Identifier: MIT */
+#include "stc_usb_service.h"
+void (*stc_usb_service)(void) STC_REENTRANT;