+++ libraries/CAN/src/Arduino_CAN.h
@@ -1,0 +1,5 @@
+/* SPDX-License-Identifier: MIT */
+#ifndef STC_ARDUINO_CAN_H
+#define STC_ARDUINO_CAN_H
+#include "CAN.h"
+#endif