+++ libraries/HID/src/HID.cpp
@@ -1,0 +1,4 @@
+/* SPDX-License-Identifier: MIT */
+#include "HID.h"
+HID_ &HID() { static HID_ instance; return instance; }
+USBDeviceClass USBDevice;