+++ libraries/HID/src/USB.h
@@ -1,0 +1,5 @@
+/* SPDX-License-Identifier: MIT */
+#ifndef STC_USB_H
+#define STC_USB_H
+#include "HID.h"
+#endif